Revert chef version to avoid issues with pg gem
The pg-ruby recipe contains a retry path that handles incompatible SSL
versions but newer versions of the chef_gem recipe break it.  By moving
to an older version of chef, we can avoid the problem.  Given we don't
really have any hard dependencies on chef versions, this is a quick way
to get things working again.

Another issue that's been addressed is the conflict between the periodic
update configuration introduced in base.sh and that managed by the
cookbooks.  Removing the code from `base.sh` seems to resolve this.