Force non-interactive kernel upgrade
Updating to the current 3.13.0-35.62 kernel is producing a 'conflict' in
/etc/menu/grub.lst, causing the unattended upgrade to fail.

A post on askubuntu (http://askubuntu.com/a/262445) proposed setting
UCF_FORCE_CONFFOLD to avoid the interactive prompt. Given that we
overwrite the grub.lst in a later stage (scripts/aws-grub.sh), it
doesn't matter if we keep the old or choose the new grub.lst entries,
so UCF_FORCE_CONFFOLD was arbitrarily chosen (as opposed to
UCF_FORCE_CONFFNEW).

DEBIAN_FRONTEND=noninteractive is added as an extra measure.