Add downloaded bosh lite stemcell to gitignore
The process of installing a bosh-lite + cf requires a bosh-lite stemcell to be downloaded. At the end of the process, my local git repository reported some changes:

```bash
± dlr |master ?:1 ✗| → git status
On branch master
Your branch is up-to-date with 'origin/master'.
Untracked files:
  (use "git add <file>..." to include in what will be committed)

        latest-bosh-lite-stemcell-warden.tgz

nothing added to commit but untracked files present (use "git add" to track)
```

This PR adds the downloaded *latest-bosh-lite-stemcell-warden.tgz* to gitignore.